# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "default": [],
            "items": {},
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ExecutionSettings": {
        "additionalProperties": false,
        "properties": {
          "max_devices": {
            "default": 4,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Max Devices",
            "type": "integer"
          },
          "min_devices": {
            "default": 1,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Min Devices",
            "type": "integer"
          },
          "start_policy": {
            "$ref": "#/components/schemas/StartPolicy",
            "default": "wait_for_min_devices"
          }
        },
        "title": "ExecutionSettings",
        "type": "object"
      },
...
      "GaussianNBParams": {
        "additionalProperties": false,
        "properties": {
          "feature_columns": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Feature Columns",
            "type": "array"
          },
          "missing_values": {
            "$ref": "#/components/schemas/MissingValuePolicy",
            "default": "drop_rows"
          },
          "split_seed": {
            "default": 42,
            "title": "Split Seed",
            "type": "integer"
          },
          "target_column": {
            "minLength": 1,
            "title": "Target Column",
            "type": "string"
          },
          "test_fraction": {
            "default": 0.2,
            "maximum": 0.5,
            "minimum": 0.05,
            "title": "Test Fraction",
            "type": "number"
          },
          "var_smoothing": {
            "default": 1e-09,
            "minimum": 0.0,
            "title": "Var Smoothing",
            "type": "number"
          }
        },
        "required": [
          "target_column",
          "feature_columns"
        ],
        "title": "GaussianNBParams",
        "type": "object"
      },
...
      "LinearRidgeParams": {
        "additionalProperties": false,
        "properties": {
          "alpha": {
            "default": 1.0,
            "minimum": 0.0,
            "title": "Alpha",
            "type": "number"
          },
          "feature_columns": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Feature Columns",
            "type": "array"
          },
          "missing_values": {
            "$ref": "#/components/schemas/MissingValuePolicy",
            "default": "drop_rows"
          },
          "split_seed": {
            "default": 42,
            "title": "Split Seed",
            "type": "integer"
          },
          "target_column": {
            "minLength": 1,
            "title": "Target Column",
            "type": "string"
          },
          "test_fraction": {
            "default": 0.2,
            "maximum": 0.5,
            "minimum": 0.05,
            "title": "Test Fraction",
            "type": "number"
          }
        },
        "required": [
          "target_column",
          "feature_columns"
        ],
        "title": "LinearRidgeParams",
        "type": "object"
      },
...
      "MissingValuePolicy": {
        "enum": [
          "drop_rows",
          "reject"
        ],
        "title": "MissingValuePolicy",
        "type": "string"
      },
...
      "StartPolicy": {
        "enum": [
          "wait_for_min_devices"
        ],
        "title": "StartPolicy",
        "type": "string"
      },
      "TaskManifest": {
        "additionalProperties": false,
        "properties": {
          "dataset_id": {
            "pattern": "^ds_",
            "title": "Dataset Id",
            "type": "string"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionSettings",
            "default": {
              "max_devices": 4,
              "min_devices": 1,
              "start_policy": "wait_for_min_devices"
            }
          },
          "kernel_version": {
            "const": "1",
            "default": "1",
            "title": "Kernel Version",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GaussianNBParams"
              },
              {
                "$ref": "#/components/schemas/LinearRidgeParams"
              }
            ],
            "title": "Params"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          }
        },
        "required": [
          "name",
          "task_type",
          "dataset_id",
          "params"
        ],
        "title": "TaskManifest",
        "type": "object"
      },
      "TaskOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "execution": {
            "additionalProperties": true,
            "title": "Execution",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kernel_version": {
            "title": "Kernel Version",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          },
          "prepared": {
            "additionalProperties": true,
            "title": "Prepared",
            "type": "object"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Status History",
            "type": "array"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "validation": {
            "additionalProperties": true,
            "title": "Validation",
            "type": "object"
          },
          "verification_policy": {
            "additionalProperties": true,
            "title": "Verification Policy",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "task_type",
          "kernel_version",
          "dataset_id",
          "params",
          "execution",
          "validation",
          "prepared",
          "status",
          "status_history",
          "verification_policy",
          "created_at"
        ],
        "title": "TaskOut",
        "type": "object"
      },
...
      "TaskType": {
        "enum": [
          "gaussian_nb_train",
          "linear_ridge_train"
        ],
        "title": "TaskType",
        "type": "string"
      },
      "TaskTypeInfo": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "kernel_version": {
            "title": "Kernel Version",
            "type": "string"
          },
          "params_schema": {
            "additionalProperties": true,
            "title": "Params Schema",
            "type": "object"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          }
        },
        "required": [
          "task_type",
          "kernel_version",
          "description",
          "params_schema"
        ],
        "title": "TaskTypeInfo",
        "type": "object"
      },
...
      "ValidationResult": {
        "properties": {
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "plan_preview": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Populated from P5 (device-aware planner).",
            "title": "Plan Preview"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "errors",
          "warnings",
          "summary"
        ],
        "title": "ValidationResult",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ProofNet API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/task-types": {
      "get": {
        "operationId": "task_types_api_v1_task_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskTypeInfo"
                  },
                  "title": "Response Task Types Api V1 Task Types Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Task Types",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskOut"
                  },
                  "title": "Response List Tasks Api V1 Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskManifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/validate": {
      "post": {
        "operationId": "validate_task_api_v1_tasks_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskManifest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Validate Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}